services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.27.1
aiokafka==0.10.0
redis==5.0.8
pydantic==2.8.2
PyJWT==2.8.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.3.2
httpx==0.27.2
